+++ src/lib/constants/sections.ts
@@ -94,7 +94,6 @@ export const SCHEMA_SECTIONS = {
     // @ts-expect-error temporary route for testing
     dataSetWIP: {
         name: SchemaName.dataSet,
-        routeName: 'dataSetsWip',
         namePlural: 'dataSets',
         title: i18n.t('Data set'),
         titlePlural: i18n.t('Data sets'),